Refactor native/cpu/zmath.h
